requirements.txt
aniso8601==7.0.0
asgiref==3.10.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==0.5.0
Django==4.2.26
django-debug-toolbar==6.1.0
django-import-export==3.3.8
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
django-rosetta==0.9.9
MarkupPy==1.18
odfpy==1.4.1
openpyxl==3.1.5
pillow==12.0.0
promise==2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
Rx==1.6.1
singledispatch==4.1.2
six==1.17.0
sqlparse==0.5.3
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.15.0
whitenoise==5.2.0
xlrd==2.0.2
xlwt==1.3.0
